refactor: extract WorkspaceExplorer from the workspace file dock
Move the useWorkspacePreview hook, all refresh/auto-refresh logic, and the
tree+preview body into a new presentation-agnostic WorkspaceExplorer component.
WorkspaceFilesPanel keeps only the dock chrome (aside, resize seam, Escape
handler). WorkspaceExplorer accepts headerActions and showFullPageLink slots so
each wrapper can inject context-specific controls without the explorer knowing
about them.

Move container-type: inline-size from .wp-dock to the new .wp-explorer so the
@container tree/preview toggle fires off the explorer width in both the dock
and the upcoming full-page context.
